Fix documentation describing reserved error codes.
Prior to this commit, the documentation gave conflicting
information about which codes are reserved for Maelstrom's use
and which are available for application-specific purposes.
Due to a typo, the two categories had overlapping ranges.

This commit consistently specifies that codes 999 and below are
reserved, while codes 1000 and above are free for application use.